fix(admin): Update SelectCohort design
