+++ admin/src/scenes/injep-export/index.tsx
@@ -2,8 +2,9 @@ import React, { useState } from "react";
 import { toastr } from "react-redux-toastr";
 import * as FileSaver from "file-saver";
 import { useDispatch, useSelector } from "react-redux";
-import { HiHome, HiOutlineCalendar } from "react-icons/hi";
+import { HiOutlineCalendar } from "react-icons/hi";
 import plausibleEvent from "@/services/plausible";
+import cx from "classnames";
 
 import { translate } from "snu-lib";
 import dayjs from "dayjs";
@@ -15,14 +16,15 @@ import SelectCohort from "@/components/cohorts/SelectCohort";
 
 import ExportBox from "./components/ExportBox";
 import DatePicker from "./components/DatePicker";
+import { CohortState } from "@/redux/cohorts/reducer";
 
-const exportDateKeys = ["cohesionCenters", "youngsBeforeSession", "youngsAfterSession"];
+const exportDateKeys = ["youngsBeforeSession", "youngsAfterSession"];
 
-const DSNJExport = () => {
+const INJEPExport = () => {
   const dispatch = useDispatch();
-  const cohortList = useSelector((state) => state.Cohorts);
+  const cohortList = useSelector((state: CohortState) => state.Cohorts);
 
-  const [currentCohort, setCurrentCohort] = useState(cohortList[0]);
+  const [currentCohort, setCurrentCohort] = useState<any>(cohortList?.[0]);
 
   const [isLDownloadingByKey, setDownloadingByKey] = useState({});
   const [isModalConfirmOpenByKey, setIsModalConfirmOpenByKey] = useState({});
@@ -30,9 +32,9 @@ const DSNJExport = () => {
   const [currentKey, setCurrentKey] = useState(exportDateKeys[0]);
   const [newExportDate, setNewExportDate] = useState();
 
-  useDocumentTitle("Export DSNJ");
+  useDocumentTitle("Export INJEP");
   const todayPlusOneDay = dayjs().add(1, "day").toDate();
-  const threeMonthsAfterCohortDateEnd = dayjs(currentCohort.dateEnd).add(3, "month").toDate();
+  const oneMonthAfterCohortDateEnd = dayjs(currentCohort.dateEnd).add(1, "month").toDate();
 
   const getExportAvailableUntilDate = (date) => {
     if (!date) return null;
@@ -41,8 +43,6 @@ const DSNJExport = () => {
 
   const translateKey = (key) => {
     switch (key) {
-      case "cohesionCenters":
-        return "centres";
       case "youngsBeforeSession":
         return "volontaires affectés et sur liste complémentaire";
       case "youngsAfterSession":
@@ -53,20 +53,20 @@ const DSNJExport = () => {
   };
 
   const generateText = (key, date) => {
-    const isNewExport = currentCohort?.dsnjExportDates?.[key] === undefined;
+    const isNewExport = currentCohort?.injepExportDates?.[key] === undefined;
     const translatedKey = translateKey(isNewExport ? key : currentKey);
     const formattedDate = dayjs(isNewExport ? date : newExportDate).format("DD/MM/YYYY");
 
     return (
-      <p className={!isNewExport && "text-red-600 text-left mt-4"}>
+      <p className={cx({ "text-red-600 text-left mt-4": !isNewExport })}>
         {isNewExport ? (
           <>
-            Vous vous apprêtez à rendre disponible à la DSNJ l'export des <span className="font-bold">{translatedKey}</span> le <span className="font-bold">{formattedDate}</span>.
+            Vous vous apprêtez à rendre disponible à l'INJEP l'export des <span className="font-bold">{translatedKey}</span> le <span className="font-bold">{formattedDate}</span>.
           </>
         ) : (
           <>
             <span className="font-bold">Attention : </span>un nouvel export des <span className="font-bold">{translatedKey}</span> sera généré le{" "}
-            <span className="font-bold">{formattedDate}</span> avec les informations mises à jour. L’ancien export ne sera plus téléchargeable par la DSNJ :{" "}
+            <span className="font-bold">{formattedDate}</span> avec les informations mises à jour. L’ancien export ne sera plus téléchargeable par l'INJEP :{" "}
             <span className="underline">pensez à les prévenir qu’ils doivent à nouveau télécharger cet export !</span>
           </>
         )}
@@ -76,17 +76,17 @@ const DSNJExport = () => {
 
   const handleUpdateDate = async (key, date) => {
     setIsModalConfirmOpenByKey({ ...isModalConfirmOpenByKey, [key]: false });
-    const { ok, code, data: updatedCohort } = await api.put(`/cohort/${currentCohort._id}/export-dsnj/${key}`, { date: dayjs(date).format("YYYY-MM-DD") });
+    const { ok, code, data: updatedCohort } = await api.put(`/cohort/${currentCohort._id}/export-injep/${key}`, { date: dayjs(date).format("YYYY-MM-DD") });
     if (!ok) return toastr.error("Une erreur est survenue lors de l'enregistrement de la date d'export", translate(code));
     dispatch({ type: COHORTS_ACTIONS.UPDATE_COHORT, payload: updatedCohort });
     setCurrentCohort(updatedCohort);
   };
 
   const handleDownload = async (key) => {
-    plausibleEvent(`Export/DSNJ - Exporter ${key}`);
+    plausibleEvent(`Export/INJEP - Exporter ${key}`);
     try {
       setDownloadingByKey({ ...isLDownloadingByKey, [key]: true });
-      const file = await api.get(`/cohort/${currentCohort._id}/export-dsnj/${key}`);
+      const file = await api.get(`/cohort/${currentCohort._id}/export-injep/${key}`);
       FileSaver.saveAs(new Blob([new Uint8Array(file.data.data)], { type: file.mimeType }), file.fileName);
     } catch (e) {
       toastr.error("Oups, une erreur est survenue pendant le téléchagement", "");
@@ -108,8 +108,9 @@ const DSNJExport = () => {
     <>
       <Page>
         <Header
-          title="Données centres & volontaires (DSNJ)"
-          breadcrumb={[{ title: "Séjours" }, { title: "Export DSNJ" }]}
+          title="Données centres & volontaires (INJEP)"
+          breadcrumb={[{ title: "Séjours" }, { title: "Export INJEP" }]}
+          // @ts-ignore
           actions={
             <SelectCohort
               cohort={currentCohort.name}
@@ -121,28 +122,22 @@ const DSNJExport = () => {
         />
         <div className="flex gap-4">
           <ExportBox
-            title="Liste des centres"
-            availableFrom={currentCohort?.dsnjExportDates?.[exportDateKeys[0]]}
-            availableUntil={getExportAvailableUntilDate(currentCohort?.dsnjExportDates?.[exportDateKeys[0]])}
+            title="Liste des volontaires au début du séjour (J+2)"
+            availableFrom={currentCohort?.injepExportDates?.[exportDateKeys[0]]}
+            // @ts-ignore
+            availableUntil={getExportAvailableUntilDate(currentCohort?.injepExportDates?.[exportDateKeys[0]])}
             onClick={() => handleClick(exportDateKeys[0])}
             onDownload={() => handleDownload(exportDateKeys[0])}
             isDownloading={!!isLDownloadingByKey[exportDateKeys[0]]}
           />
           <ExportBox
-            title="Liste des volontaires affectés et sur liste complémentaire"
-            availableFrom={currentCohort?.dsnjExportDates?.[exportDateKeys[1]]}
-            availableUntil={getExportAvailableUntilDate(currentCohort?.dsnjExportDates?.[exportDateKeys[1]])}
+            title="Liste des volontaires après le séjour (J+6)"
+            availableFrom={currentCohort?.injepExportDates?.[exportDateKeys[1]]}
+            // @ts-ignore
+            availableUntil={getExportAvailableUntilDate(currentCohort?.injepExportDates?.[exportDateKeys[1]])}
             onClick={() => handleClick(exportDateKeys[1])}
             onDownload={() => handleDownload(exportDateKeys[1])}
             isDownloading={!!isLDownloadingByKey[exportDateKeys[1]]}
-          />
-          <ExportBox
-            title="Liste des volontaires après le séjour"
-            availableFrom={currentCohort?.dsnjExportDates?.[exportDateKeys[2]]}
-            availableUntil={getExportAvailableUntilDate(currentCohort?.dsnjExportDates?.[exportDateKeys[2]])}
-            onClick={() => handleClick(exportDateKeys[2])}
-            onDownload={() => handleDownload(exportDateKeys[2])}
-            isDownloading={!!isLDownloadingByKey[exportDateKeys[2]]}
           />
         </div>
         <DatePicker
@@ -150,8 +145,10 @@ const DSNJExport = () => {
           onChangeDate={(key, date) => handleChangeDate(key, date)}
           isOpen={isDatePickerOpenByKey[currentKey]}
           onClose={() => setIsDatePickerOpenByKey({ ...isDatePickerOpenByKey, [currentKey]: false })}
+          // @ts-ignore
           minDate={todayPlusOneDay}
-          maxDate={threeMonthsAfterCohortDateEnd}
+          // @ts-ignore
+          maxDate={oneMonthAfterCohortDateEnd}
         />
         <ModalConfirmation
           isOpen={!!isModalConfirmOpenByKey[currentKey]}
@@ -164,7 +161,7 @@ const DSNJExport = () => {
               <HiOutlineCalendar size={24} />
             </div>
           }
-          title="Modification de la date de mise à disponibilité de l’export DSNJ"
+          title="Modification de la date de mise à disponibilité de l’export INJEP"
           text={generateText(currentKey, newExportDate)}
           actions={[
             { title: "Annuler", isCancel: true },
@@ -176,4 +173,4 @@ const DSNJExport = () => {
   );
 };
 
-export default DSNJExport;
+export default INJEPExport;
